requirements.txt
langchain>=0.2.9         # required for compatibility with langchain-chroma
langchain-openai>=0.0.5
langchain-chroma>=0.0.3  # new standalone Chroma package (replaces langchain_community Chroma)
openai>=1.3.5
chromadb>=0.4.24
beautifulsoup4>=4.12.2
nest_asyncio>=1.5.8
tqdm>=4.66.1
python-dotenv>=1.0.1
